Move Add workspace to data presenter
